+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Rename `maybeOf` to `maybe`, `eitherOf` to `either`, `promiseOf` to `promise`, `tupleOf` to `tuple`, `arrayFrom` to `array`.
 - Makes inner values if `Wrapper`, `Maybe` and `Either` types as ECMAScript private fields.
+- `array` function can accept now plain values (along with `ArrayLike` and `Iterable` objects) and will add it to resulting array.
 
 ## [0.9.1] - 2020-11-29